res.sendStatus(401);             return;

+++ src/routes/devises_router.ts
@@ -10,12 +10,8 @@ devicesRouter.get('/', async (req, res)=>{
         const checkToken = await jwtService.verifyUserIdByRefreshToken(cookie)
         const time = await jwtService.getLastActiveDateFromRefreshToken(cookie)
         const session = await devicesService.findLastActiveDate(time)
-        if(!checkToken || !session){
-            res.sendStatus(401);
-            return;
-        }
-        const sessions:Array<DeviceViewType>|null = await devicesService.getAllCurrentSessions(checkToken.userId)
-        if(sessions){
+        const sessions:Array<DeviceViewType>|null = await devicesService.getAllCurrentSessions(checkToken?.userId)
+        if(sessions && session){
             res.status(200).send(sessions)
             return;
         }
@@ -36,13 +32,14 @@ devicesRouter.delete('/', async (req, res)=>{
         const checkToken = await jwtService.verifyUserIdByRefreshToken(cookie)
         const time = await jwtService.getLastActiveDateFromRefreshToken(cookie)
         const session = await devicesService.findLastActiveDate(time)
-        if(!checkToken || !session){
-            res.sendStatus(401);
-            return;
-        }
-        else{
+        if(checkToken && session){
             await devicesService.deleteAllSession(session.userId,session.deviceId)
             res.sendStatus(204);
+            return;
+
+        }
+        else{
+            res.sendStatus(401);
             return;
         }
     }
